DealiTextInput 개선
