+++ DealiDesignSystemSampleApp/DealiDesignSystemSampleApp/ViewControllers/Sub/TextInputViewController.swift
@@ -68,7 +68,6 @@ final class TextInputViewController: UIViewController {
             let button = DealiControl.btnOutlineMedium01()
             button.title = "Default"
             $0.actionButton = button
-            $0.inputRightViewType = .clear
         }.snp.makeConstraints {
             $0.left.right.equalToSuperview()
         }
